Extracted DefaultRoute
