+++ Tests/Unit/PathMask.phpt
@@ -1,0 +1,39 @@
+<?php
+declare(strict_types = 1);
+/**
+ * @testCase
+ * @phpVersion > 7.1
+ */
+namespace Klapuch\Routing\Unit;
+
+use Klapuch\Routing;
+use Klapuch\Uri;
+use Tester;
+use Tester\Assert;
+
+require __DIR__ . '/../bootstrap.php';
+
+final class PathMask extends Tester\TestCase {
+	public function testNamesByPosition() {
+		Assert::same(
+			['name' => 'dom', 'position' => 'developer'],
+			(new Routing\PathMask(
+				'/books/{name}/{position}',
+				new Uri\FakeUri(null, '/books/dom/developer', [])
+			))->parameters()
+		);
+	}
+
+	public function testRemovingRegexParts() {
+		Assert::same(
+			['foo' => '10', 'adjective' => 'cool', 'word' => 'bar'],
+			(new Routing\PathMask(
+				'/books/{foo \d+}/{adjective \w+}/{word}',
+				new Uri\FakeUri(null, '/books/10/cool/bar', [])
+			))->parameters()
+		);
+	}
+}
+
+
+(new PathMask())->run();
